added options to deployment and reset scripts

+++ infra/Site_Reset.sh
@@ -1,12 +1,23 @@
 #!/bin/bash
-# comment/uncomment lines as needed
-# stops and removes the site's containers (--rm flag during run command deletes container upon exiting)...
-#... and erases the node web server image (if needed for development)
-# stop and remove node web server container
-echo 'stopping and removing the following container...'
-docker stop web_server
+
+# ask what mode the web application is running in
+read -p "what mode was the web application started in, production or development?(enter p or d): " mode
+# following code is executed for both modes
 # stop and remove postgreSQL container
 echo 'stopping and removing the following container...'
 docker stop database
-# remove node web server image
-docker rmi michaelt23/web:server
+# if production mode
+if [ "$mode" == "p" ]; then
+    # stop and remove node web server container
+    echo 'stopping and removing the following container...'
+    docker stop web_server
+    # remove node web server image
+    docker rmi michaelt23/web:server
+    echo 'web application reset complete'
+# if development mode (just input validation currently)
+elif [ "$mode" == "d" ]; then
+    echo 'web application reset complete'
+# invalid mode
+else
+    echo 'invalid input'
+fi
